feat: check & request user permissions for the FS
If the tab is stale on Chrome the user needs to give permissions again
for access to the Directory Handle to work.

At this stage both Chrome & Electron clients work with the filesystem
API. With the only difference that on Electron on every page reload
the Directory Handle PermissionState is "prompt" so it requires a user
action to grant proper permissions again.

More on
* https://stackoverflow.com/a/69897694 &
* https://developer.chrome.com/blog/persistent-permissions-for-the-file-system-access-api